Remove LeaderAndRequest
