+++ metadata/src/main/java/org/apache/kafka/metadata/PartitionRegistration.java
@@ -18,12 +18,12 @@
 package org.apache.kafka.metadata;
 
 import org.apache.kafka.common.DirectoryId;
+import org.apache.kafka.common.PartitionState;
 import org.apache.kafka.common.TopicPartition;
 import org.apache.kafka.common.Uuid;
 import org.apache.kafka.common.errors.InvalidReplicaDirectoriesException;
 import org.apache.kafka.common.metadata.PartitionChangeRecord;
 import org.apache.kafka.common.metadata.PartitionRecord;
-import org.apache.kafka.common.requests.LeaderAndIsrRequest;
 import org.apache.kafka.image.writer.ImageWriterOptions;
 import org.apache.kafka.server.common.ApiMessageAndVersion;
 
@@ -410,9 +410,9 @@ public class PartitionRegistration {
         return new ApiMessageAndVersion(record, options.metadataVersion().partitionRecordVersion());
     }
 
-    public LeaderAndIsrRequest.PartitionState toLeaderAndIsrPartitionState(TopicPartition tp,
+    public PartitionState toLeaderAndIsrPartitionState(TopicPartition tp,
                                                                            boolean isNew) {
-        return new LeaderAndIsrRequest.PartitionState().
+        return new PartitionState().
             setTopicName(tp.topic()).
             setPartitionIndex(tp.partition()).
             setControllerEpoch(-1).
